feat: add leader election, game mode (coop/race), mode-specific rendering
- Add GameMode enum (COOP/RACE) in common
- Server tracks session leader (first connected, promoted on disconnect)
- Leader can toggle game mode (TAB) and start game (G) in lobby
- GameState broadcasts game_mode and leader_id to all clients
- Race mode: no player collisions, no checkpoints in generated maps
- Bold font for leader name display
- Separate rendering files per mode (HudCoop/HudRace, LevelResultsCoop/LevelResultsRace, SessionResultsCoop/SessionResultsRace)
- Offline mode defaults to RACE
- Lobby options panel shows mode, leader, and controls
- Fix pre-existing build errors (std::fmodf→fmod, std::fabsf→fabs, std::sqrtf→sqrt)

+++ src/client/SoundPool.cpp
@@ -64,7 +64,7 @@ void SoundPool::PlayAt(float source_x, float source_y,
 
     const float dx   = source_x - listener_x;
     const float dy   = source_y - listener_y;
-    const float dist = std::sqrtf(dx * dx + dy * dy);
+    const float dist = std::sqrt(dx * dx + dy * dy);
 
     // Cull sounds beyond max_dist (saves processing audio for far-off players).
     if (dist >= max_dist) return;
